TFC Config + Audit event refactor to fix performance tests.
