+++ test/models/input/TFCEligibilityInputSpec.scala
@@ -112,91 +112,6 @@ class TFCEligibilityInputSpec extends CCConfigSpec with FakeCCEligibilityApplica
           TFCDisability(disabled = true, severelyDisabled = true))
         child.isDisabled shouldBe true
       }
-
-      "determine childs Birthday(11th or 16th) where child is not disable" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2005-08-27", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability()) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val child11Birthday = child.getChildBirthday(current, "england")
-        LocalDate.fromDateFields(child11Birthday) shouldBe LocalDate.parse("2016-08-27", formatter)
-      }
-
-      "determine childs Birthday (11th or 16th) where child is disabled" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2000-09-27", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability(disabled = true)) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val child11Birthday = child.getChildBirthday(current, "england")
-        LocalDate.fromDateFields(child11Birthday) shouldBe LocalDate.parse("2016-09-27", formatter)
-      }
-
-      "determine end of Week following 1st September Date where child is not disabled" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2005-08-27", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability()) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val endWeek1stOfSeptemberDate = child.endWeek1stOfSeptemberDate(current, "england")
-        LocalDate.fromDateFields(endWeek1stOfSeptemberDate) shouldBe LocalDate.parse("2016-09-04", formatter)
-      }
-
-      "determine end of Week following 1st September Date where child is not disabled, dob after 1st Sept" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2005-09-12", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability()) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val endWeek1stOfSeptemberDate = child.endWeek1stOfSeptemberDate(current, "england")
-        LocalDate.fromDateFields(endWeek1stOfSeptemberDate) shouldBe LocalDate.parse("2017-09-03", formatter)
-      }
-
-      "determine end of Week following 1st September Date where child is disabled" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2005-08-27", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability(disabled = true)) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val endWeek1stOfSeptemberDate = child.endWeek1stOfSeptemberDate(current, "england")
-        LocalDate.fromDateFields(endWeek1stOfSeptemberDate) shouldBe LocalDate.parse("2021-09-05", formatter)
-      }
-
-      "determine end of Week following 1st September Date where child is disabled, dob after 1st Sept" in {
-        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-        val dateOfBirth = LocalDate.parse("2005-09-12", formatter)
-        val current = LocalDate.parse("2017-08-01", formatter)
-        val child = new TFCChild(
-          id = 0,
-          childcareCost = BigDecimal(300.00),
-          childcareCostPeriod = Periods.Monthly,
-          dob = dateOfBirth,
-          disability = TFCDisability(disabled = true)) { override lazy val tFCConfig: TFCConfig = tfcConfig }
-        val endWeek1stOfSeptemberDate = child.endWeek1stOfSeptemberDate(current, "england")
-        LocalDate.fromDateFields(endWeek1stOfSeptemberDate) shouldBe LocalDate.parse("2022-09-04", formatter)
-      }
-
     }
